pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multiple-env-data-consistency"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = []

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
    "pytest-html",
    "allure-pytest",
    "pyyaml",
    "requests",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--html=reports/report.html --self-contained-html --alluredir=reports/allure-results"
log_cli = true
log_cli_level = "INFO"

[tool.setuptools.packages.find]
where = ["src"]
